refactor(llm): 模型层下沉 tiangong-llm，消除 ModelProviderConfig
把模型层类型从 core 下沉到 tiangong-llm，形成闭环：core 只通过 llm
提供的类型发起请求，自身不再持有模型连接配置类型。同步消除
ModelProviderConfig 中间层与 env-var fallback。

迁移到 tiangong-llm：
- ModelEndpoint（新 endpoint.rs，端点描述）
- ModelsConfig + ModelCapability/RoutingSlot/ProviderConfig/ModelEntry/
  ResolvedModel/ProviderReferences + 纯路由方法（新 models_config.rs）
- SingleProviderClient + ModelClient trait + ModelRequest/Response/StreamChunk
  + OnRetryCallback + 全部 provider 构建辅助函数（新 provider_client.rs）

消除：
- ModelProviderConfig 整类型（client 改吃 ModelEndpoint）
- ModelProviderConfig::from_env() + env-var fallback（loader/lifecycle）
- ModelsConfig::from_legacy/to_chat_provider_config/to_lite_provider_config/
  from_llm_config（往返转换层）
- parse_timeout_ms/parse_function_timeout_ms（timeout 已是 u64，直接用）

core 改为 re-export：model.rs / models_config.rs / core_config.rs
均从 llm import，外部路径 tiangong_core::model::SingleProviderClient 等
保持兼容。SingleProviderClient::new 签名 ModelProviderConfig → ModelEndpoint。

调用点更新：app-state（model_config → model_endpoint，4 facade + store）、
build_engine_from_config（直接吃 config.llm.chat）、entry model/doctor、
analyze-attachment plugin、tauri commands。

chat_is_multimodal 行为保持：build_engine_from_config 重建最小 ModelsConfig
（仅注册 Chat 能力，与旧 from_llm_config 一致），GUI/Server 路径仍注入完整
ModelsConfig。

验证：cargo check --workspace 全绿零警告；tiangong-llm 40 测试通过
（迁移主体，含迁入的 client 测试）。
