+++ crates/tiangong-llm/Cargo.toml
@@ -15,7 +15,8 @@ thiserror = { workspace = true }
 async-trait = { workspace = true }
 async-openai = { workspace = true, features = ["byot", "chat-completion", "model"] }
 backoff = { workspace = true }
-tokio = { workspace = true, features = ["rt", "time"] }
+tokio = { workspace = true, features = ["rt", "rt-multi-thread", "time", "sync"] }
 reqwest = { workspace = true, default-features = false, features = ["json", "rustls"] }
 futures-util = { workspace = true }
 tracing = { workspace = true }
+base64 = { workspace = true }
